+++ src/components/Clock/Clock.tsx
@@ -5,7 +5,7 @@ import { useCounter } from 'hooks/useCounter'
 import { Heading2 } from '@entur/typography'
 import classes from './Clock.module.scss'
 
-function Clock({ className }: { className?: string }): JSX.Element {
+function Clock({ className }: { className?: string }){
     // This insures that the Clock-component is updated once every second.
     useCounter()
