+++ src/routes/Dashboard/BusStop/components/BusStopTableRow/BusStopTableRow.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import { Departure } from 'hooks/use-stop-place-with-estimated-calls/departure'
 import { SituationInfo } from 'components/SituationInfo/SituationInfo'
+import { Departure } from 'src/types'
 import { DataCell, TableRow } from '@entur/table'
 import classes from './BusStopTableRow.module.scss'
