+++ src/routes/Dashboard/Map/DepartureTag/DepartureTag.tsx
@@ -1,12 +1,11 @@
 import React, { useMemo } from 'react'
 import { format, isToday } from 'date-fns'
 import { nb } from 'date-fns/locale'
-import { IconColorType } from 'src/types'
+import { Departure, IconColorType } from 'src/types'
 import { getIconColor } from 'utils/icon'
 import { TransportModeIcon } from 'components/TransportModeIcon/TransportModeIcon'
 import { useStopPlaceWithEstimatedCalls } from 'hooks/use-stop-place-with-estimated-calls/useStopPlaceWithEstimatedCalls'
 import {
-    Departure,
     filterHidden,
     toDeparture,
 } from 'hooks/use-stop-place-with-estimated-calls/departure'
