+++ src/routes/Dashboard/Responsive/ResponsiveDepartureTile/ResponsiveTable/ResponsiveTableRow/ResponsiveTableRow.tsx
@@ -1,16 +1,13 @@
 import React from 'react'
 import { SituationInfo } from 'components/SituationInfo/SituationInfo'
 import { TransportModeIcon } from 'components/TransportModeIcon/TransportModeIcon'
-import { Departure } from 'hooks/use-stop-place-with-estimated-calls/departure'
-import { IconColorType } from 'src/types'
+import { Departure, IconColorType } from 'src/types'
 import { getIconColor } from 'utils/icon'
 import { isDarkOrDefaultTheme } from 'utils/utils'
 import { useSettings } from 'settings/SettingsProvider'
 import classes from './ResponsiveTableRow.module.scss'
 
-const ResponsiveTableRow: React.FC<{ departure: Departure }> = ({
-    departure,
-}) => {
+function ResponsiveTableRow({ departure }: { departure: Departure }) {
     const [settings] = useSettings()
     const iconColor = getIconColor(
         departure.transportMode,
