+++ src/routes/Dashboard/Timeline/TimelineDeparture/TimelineDeparture.tsx
@@ -2,8 +2,7 @@ import React from 'react'
 import { timeUntil } from 'utils/time'
 import { TransportModeIcon } from 'components/TransportModeIcon/TransportModeIcon'
 import { useCounter } from 'hooks/useCounter'
-import { Departure } from 'hooks/use-stop-place-with-estimated-calls/departure'
-import { IconColorType } from 'src/types'
+import { Departure, IconColorType } from 'src/types'
 import { competitorPosition } from '../utils'
 import classes from './TimelineDeparture.module.scss'
