+++ src/utils/utils.ts
@@ -1,9 +1,8 @@
 import { TranslatedString } from 'graphql-generated/mobility-v2'
-import { EnturLogoStyle, Theme, TileSubLabel } from 'src/types'
+import { Departure, EnturLogoStyle, Theme, TileSubLabel } from 'src/types'
 import EnturBlack from 'assets/logos/EnturBlack.svg'
 import EnturWhite from 'assets/logos/EnturWhite.svg'
 import EnturContrast from 'assets/logos/EnturContrast.svg'
-import { Departure } from 'hooks/use-stop-place-with-estimated-calls/departure'
 
 export function createTileSubLabel({
     situations,
